+++ HISTORY.md
@@ -12,14 +12,14 @@ TODO
 0.2.0.dev0
 ----------
 
-* move base location from /api/ to /smarter-api/
+* move base location from `/api/` to `/smarter-api/`
 * document API with `flasgger` [#6](https://github.com/cnr-ibba/SMARTER-backend/issues/6)
 * add endpoint for supported assemblies variants [#11](https://github.com/cnr-ibba/SMARTER-backend/issues/11)
 * add info endpoint
 * query with multiple parameters [#23](https://github.com/cnr-ibba/SMARTER-backend/issues/23)
 * return 4xx error when querying with wrong parameters [#22](https://github.com/cnr-ibba/SMARTER-backend/issues/22)
 * Support `type` argument when searching samples
-* Track version with bump2version
+* Track version with `bump2version`
 
 0.1.0 (2021-11-11)
 ------------------